Refactor policy stuff into its own file. Also lay some groundwork for a non-exclusive mode for the affinity policy
